Fixed game state. Added pallet eating
Now collecting pallets. Score is being kept in the Snake object, but not desplayed and not used to adjust difficulty. Now displaying Game Over and behaving properly when hitting walls. Still not biting. Still no looping. Need to also provide GUI functions to navigate game.

+++ grid.cpp
@@ -20,13 +20,13 @@ unsigned int Grid::GetGridSize() {
 }
 
 Cell* Grid::GetCellAt(int i, int j) {
-    if(i >= grid_size) {
+    if(i >= (int) grid_size) {
         i = grid_size - 1;
     } else if(i < 0) {
         i = 0;
     }
 
-    if(j >= grid_size) {
+    if(j >= (int) grid_size) {
         j = grid_size - 1;
     } else if(j < 0) {
         j = 0;
